+++ examples/12-symfony/config/bundles.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    // … your other bundles …
+    BugBoard\Symfony\BugBoardBundle::class => ['all' => true],
+];